fix(header): CTA follows platform signup stage instead of always "Start free"
The site header hard-coded "Start free" → /auth on every page, ignoring the
platform stage — so in a coming-soon (waitlist_only) launch it invited signups
while the rest of the site correctly said "Join the waitlist" (confirmed live
on dev.peakhour.ai).

Split Header into a thin async server wrapper that resolves the CTA from the
live catalog (getPublicCatalog → signupCta(signupMode)) and passes it to the
existing interactive shell, now `HeaderNav` (client). The button label tracks
the stage — "Start free" when open, "Join the waitlist"/"Request an invite"
pre-launch, hidden when closed. Minimal mode (legal pages) skips the fetch.
All 15 Header importers are server components, so the server wrapper is safe.
